requirements.txt
numpy
nltk
pandas
tqdm
torch
torchvision
matplotlib
scikit-learn
gdown
contractions
pyspellchecker
pyyaml